Fixes several tests that recently regressed.
My fix for the Releaser.Dispose ended up failing several tests (somehow
unnoticed).  This fixes the regression.  I admit it feels a bit hacky in
the new fix, but it works, and I don't believe it should lead to any
deadlocks.  As with TDD so far, prove it doesn't work by writing another
test. :)