Fixes failing unit test.
And hopefully simplifies the code and improves efficiency in the process.
Before this change, every posted message was re-posted everywhere possibly legal.
This caused a single message to multiply (GC pressure, possibly) and for queues to get messagse that might not matter after a Join is canceled.
With this fix, queues are immediately disconnected when a Join is canceled, so only the exactly required set of message should ever be executed by a given AsyncPump.