Fixes failing unit test.
Fixing this bug necessarily ran us up to fixing a few other quirky issues.
A very important issue is that the OnBeforeLockReleasedAsync that ReleaseAsync calls wasn't completing before OnExclusiveLockReleasedAsync was invoked (so that they can overlap) which seemed very broken and is what led to the "reentrancy" into the ReleaseAsync method that the failing unit test caught.
As part of this fix, I settled for reducing complexity by removing the asynchronous callback behavior when releasing exclusive locks. I've updated a few test so that they don't fail with the new behavior.