Fixes hang due to work stealing between pumps.
When two pumps are stacked on the main thread due to COM reentrancy during the synchronous wait, the pump at the bottom would sometimes "steal" work from the nested pump's queue. Yet because the outer pump wasn't on the top of the stack it couldn't execute that work, and now neither could the nested one because it wasn't in its queue for it to find.
The fix is to move the actual dequeue call to only take place synchronously on the main thread rather than on a threadpool thread.