Fixes failing GC pressure test.
The issue was that the MethodInfo property we were querying was actually causing additional memory allocations, which we want to avoid except in cases of actual hangs.

The fix is to store the delegate itself rather than its MethodInfo. We only query MethodInfo when a hang occurs. We also release the delegate when the task has completed to avoid memory leaks.

Since the value can transition from non-null to null, I altered the hang reporting to defend against NullReferenceExceptions.