Adds JoinableTaskContextNode to allow custom factories.
Ownership of the JoinableTaskContext singleton instance is moving to VS.
CPS still must be able to customize the behavior of CreateFactory and hook into hang reports.
JoinableTaskContextNode is a sort of satellite that allows folks to have something that feels much like a JoinableTaskContext, but still allows a singleton pattern.