Moves UnderlyingSynchronizationContext from JoinableTaskContext to JoinableTaskFactory.
Removes the JoinableTaskContext.UnderlyingSynchronizationContext
property from the public surface area. Adds it as a 'protected' property
on JoinableTaskFactory.

This retains its availability to those who need it, while hiding it from
casual consumers so they're more likely to fall into the 'pit of
success' rather than misuse the SyncContext.

As one case study: we had a VS developer believe he was using the new
way of getting to the UI thread by writing code that called:
JoinableTaskContext.UnderlyingSynchronizationContext.Send(delegate)
This is of course an anti-pattern and actually circumvents the new
system rather than using it. So hiding this property will deter such
misuse in the future.