Expose a method to set all resources to unknown state.
The root cause is in the locking service which ignores the the lock options passed when asking for a resource which was already prepared to be accesses under the same lock but a different set of options.
The current fix works around the problem by explicitly clearing the resource preparation flag before asking for the resource in another write lock where specific options are needed.
It is a complete fix but in a non-ideal fashion.