Fix crash that can occur after disposal of the AsyncReaderWriterLock.
When someone captures the SynchronizationContext applied to an UpgradeableRead or Write lock issued by an AsyncReaderWriterLock, and calls Post on it such that that call or the delegate doesn't execute until after the SynchronizationContext has been disposed, an ObjectDisposedException can be thrown on a threadpool thread such that it crashes the process.

We saw non-deterministic unit test failures in CPS as a result of this. This fix avoids the crash and reliably executes the delegate.