Avoid running the GC repeatedly on suspected hangs.
For long running tasks, we have seen the GC take a significant portion of the time as a result of forced GCs. Eventually we should get rid of them entirely. But until we get better tooling for crawling memory dumps and discarding stale async state machines, a GC still makes it easier to analyze. So we increase the timeout and only run GC once instead of in a loop to reduce its impact.