Fix ArgumentNullException thrown from JoinRelease.Dispose
When a JoinableTaskCollection is joined within a JoinableTask, if the resulting JoinRelease struct is disposed of twice, an ArgumentNullException is thrown because this.joiner has been cleared but this.joinedJobCollection was not.

It looks like a copy/paste bug, and has been present since 2012. But I recently noticed it due to some code I was writing that led to double-disposal.