Allow naming of JoinableTaskCollection instances for hang reports.
Hang reports are notoriously vague when it comes to JoinableTaskCollection nodes which are just called "Collection #1", "Collection #2" and so on. With this change, they can now be named so that folks that create them can name them after the type that created them (for example) so that correlating the DGML graph back to code is more straightforward.