Fix leaks of CancellationTokenRegistration.
JTF.MainThreadAwaiter.OnCompleted() registers a cancellation action to a cancellation token source but that cancellation token registration is not disposed. Ideally the cancellation token registration should be disposed when JTF.MainThreadAwaiter.GetResult() is called. However, there is a race condition between JTF.MainThreadAwaiter.OnCompleted() and JTF.MainThreadAwaiter.GetResult(), and JTF.MainThreadAwaiter.GetResult() could be executed on main thread even before JTF.MainThreadAwaiter.OnCompleted() registers the cancellation action on the background thread. That sounds very possible since JTF.MainThreadAwaiter.OnCompleted() posts the continuation to main thread queue and then registers the cancellation action. The fix is to acquire a lock in both methods to synchronize the operations on that cancellation token registration. A test was added to verify that the leak existed before the fix and was gone after the fix.

The ordering change in JoinableTask.cs is for the testing purpose, because the test code is hooking with PostToUnderlyingSynchronizationContextOrThreadPool() and needs to intercept the wait() code after the queue events are signaled. That ordering change looks harmless, otherwise, we need to invent a new mechanism to intercept that wait() code.