Avoid ArgumentNullException thrown on .NET 4.5
The TaskCompletionSource<T>.TrySetCanceled(CancellationToken) method only exists in .NET 4.6. When .NET 4.5 is on the box (such as in the devenv /setup scenario during initial install) this method doesn't exist and results in our calling CreateDelegate passing in a null MethodInfo.

The fix is to check for null and only try to call CreateDelegate when it's successful. The code that uses this field is already prepared for the TrySetCanceled static field to be null.