Use open delegates for AsyncLocal<T>
The closed delegates that were unique per instance of AsyncLocal<T> made creating new instances of AsyncLocal<T> quite slow. With this change, creation time is 40X faster. Access to the Value property is twice as slow (4 ms instead of 2ms for 50,000 accesses) but within acceptable range for the benefit we're seeing.