Fix race condition in AsyncLazy<T> that can lead to NullReferenceException
When the value factory completes asynchronously but very quickly afterward, the GetValueAsync method can throw a NullReferenceException because this.joinableTask is nulled out after the value factory completes.
The fix is to change the check then re-read field pattern to instead use ?., which compiles to a read once, then check and use as a local field. It's harmless to call JoinAsync on a completed JoinableTask.

Note that while the test changes were necessary to be able to reproduce the NRE (because a JTF is required), the test naturally executing does not tend to hit the problem. Instead, I had to freeze threads in the debugger while running the test.