Simplify AsyncLazy value factory cleanup with async delegate
We only recently introduced this delegate around the provided value factory, and with that came a small tax. We can make that tax work better for us by moving our cleanup code into it rather than as an extra Task object that must be allocated, along with (likely) an allocated collection to handle the multiple continuations off of the this.value Task.