Fix WaitWithoutInlining() test to return deterministic results
Before this change, it relied on its ability to call Task.Wait before the delegate Task.Run happened to start executing on the ThreadPool. If the race went south, the test would pass without having verified what it claimed.
This fix greatly increases the odds of it testing what it claims to.