Fix test stability: OnHangDetected_RunAsync_OnMainThread_BlamedMethodIsEntrypointNotBlockingMethod
The issue with
OnHangDetected_RunAsync_OnMainThread_BlamedMethodIsEntrypointNotBlockingMethod
was that hangs are reported at a very short interval in this test, and
that creates a race with the completion of the jt task in the test. If
the jt task completes as the last hang report is being written, the
entrypoint method for that task has been cleared and the report is
incomplete. This clearing of the entrypoint method behavior is by design
to avoid memory leaks. So I've adjusted the test to always look at the
first hang report.

Fixes #80