Fix race condition that can lead to a hang in test
The issue with StressMultipleGroups was a volatile read of a local
variable that made a decision and then acted on that decision later
while other threads were also manipulating the variable. I've changed
the test logic to not depend on that variable at all and to break out of
the loop under a simpler condition.

Fixes #119