Fix race condition in test TransitionToMainThreadRaisedWhenSwitchingToMainThread
If the threadpool executed the Task.Run delegate fast enough, the returned Task can actually be completed before the calling thread tries to await on it, leading to the next assertion to fail because the main thread transitions never took place.

I fix this by forcing a yield in all cases.

Fixes #44