Fix unreliable tests
Tests:
 * RunAsyncWithYieldingDelegateNestedInRunOverhead
 * RunAsyncWithNonYieldingDelegateNestedInRunOverhead

Somehow we were getting a WaitCount of 0 at the end of the test, which I can only explain by suggesting that the 1000 iteration loop on a background thread completed before the main thread even had to sleep once. This seems hardly expected, but it's the only explanation I can think of. So this commit changes the test to *ensure* it waits at least once.

Fixes #45
Fixes #46