VSTHRD010 code fix should find the inner-most node for a given span
When an ArgumentSyntax is made up entirely of an ExpressionSyntax that violates VSTHRD010, an InvalidCastException would be thrown by the code fix provider. This fixes that by always preferring the most derived node that matches the span.

If that someday fails as well, we can choose the outer-most or inner-most and then walk the tree till we find what we expect.

Fixes #270