Add Task.ConfigureAwaitRunInline extension method
This makes it easy to avoid a dependency on the UI thread or the threadpool to complete async methods that have very little work to do after a Task completes.

.ConfigureAwait(false) does *not* inline continuations when SynchronizationContext.Current != null (when the antecedent is completed).
.ConfigureAwait(true) does *not* inline continuations unless SynchronizationContext.Current is the same instance from when the continuation was created to when the antecedent completes.

Yet .ConfigureAwaitRunInline() will *always* inline the continuation so long as the TPL feels there is enough stack space available. Otherwise it will fallback to the threadpool.