Refactory the dependency tracking data structure
Pull JoinableTaskDependentNode out of JTF, so JTF collection can be treated as a different
kind of nodes, just like a normal JTF.

This is a messy change, because of all the visibility issue, and we cannot expose any API to external code, which eliminate the possibility of using 'protected' members.  I am not happy with many internal methods, but need get things to the working state first.