Overload ApplyChangeOptimistically to take operand
This allows elimination of closure allocations in many common cases.

Consider adding an item to an immutable collection. With this change,
the item to be added can be passed on the stack, making the lambda
'static', and avoiding a per-call closure heap allocation.

```c#
ImmutableList<T> list;
ThreadingTools.ApplyChangeOptimistically(
    ref list, itemToAdd, (list, item) => list.Add(item));
```

This URL shows the difference and benefit:

https://sharplab.io/#v2:EYLgZgpghgLgrgJwgZwLQFEAeAHJzkCWA9gHYBqUCBUwANigD4ACADAARMCMA3ALABQApgGYOAJjYBhAQG8BbBR1FMALGwCynABQESMNgQCUbeYrn9FlqVGzwkWrcYC8APgNsnTti0N8LigF9TBWClDjV1MR09A2NQ8ytFADkiSRs7CB0AGjYAKw83fM9vX1Cg/xCKtmAiIlprW0RMpgBWAB4aurcwOBIAY2du3r7HP0tQzvqUtMb7XX0CHNa2+ZzJof7Bth7+nV9x/gCgA=