Fix CS8600 (Converting null literal or possible null value to non-nullable type)
This diagnostic can be fixed in nearly all cases without signature
changes, so it's safe to correct this warning early in the transition to
nullable reference types and then maintain it as development proceeds.

For the specific case of Utils.IsDerivedFrom (the only signature change
in this commit), three options were available: 1) create a nullable
local variable which is initialized to the parameter 'type' and operate
on the local variable instead of the parameter; 2) redefine the
parameter as nullable but apply the [DisallowNull] attribute to it to
indicate that the initial value is not null but subsequent assignments
can set it to null; or 3) redefine the parameter as nullable and allow
null as an input. For this method, option 3 was semantically correct and
the method is internal so external callers did not need to be
specifically considered.