Suppress CS8602 using targeted assertions (Dereference of a possibly null reference)
The compiler is not always able to figure out that a value is not null
at a particular program location. For cases that are not easy to rewrite
to other forms, an assertion is added to the code as a hint to flow
analysis that a value is not null at that location.