Fix CS8604 by allowing null arguments (Possible null reference argument)
Many methods in the code are implemented to allow null arguments. In
some cases, the methods also return a Boolean value which can only have
a particular value if the parameter is not null. For these cases, the
attribute [NotNullWhen] is added to the parameter. While care must be
taken to (manually) ensure the accuracy of this attribute, the ideal
situation involves methods implemented with a simple null check at the
beginning. For example:

```
if (parameterName is null)
  return false;
```

A method starting with this block can have [NotNullWhen(true)] applied
to `parameterName`.