Fix CS8604 in analyzer test code (Possible null reference argument)
This change is made with understanding of the specific behavior of the
VerifyCodeFixAsync method used by this test. When this API gets the same
value for the source and fixedSource parameters, it automatically
validates that the code fix provider does not offer any fix for the
diagnostic(s) present in the source. This behavior matches the current
intent of the test.