Fix CS8604 by rewriting conditions (Possible null reference argument)
This commit updates code so the compiler can tell a specific local is
not null. The result is not semantically identical to the original code,
but the change is not observable because it is not possible to have an
InvocationExpressionSyntax that provides a non-null symbol that also
does not implement IMethodSymbol. Even if such a case did exist, it
would not be relevant for this analyzer that only operates on methods.