Fix logging of mismatched arguments
The call to `WriteEvent` must have the same argument types as the parameters of
the method with the `[Event]` attribute. The comparison in mscorlib doesn't
coerce between enum/int.

Prior to this change, the debugger displays a bunch of messages in the output
window:

> The parameters to the Event method do not match the parameters to the WriteEvent method. This may cause the event to be displayed incorrectly.

This is logged in System.Diagnostics.Tracing.EventSource.LogEventArgsMismatches.