Re-implement AsyncSemaphore to remove SemaphoreSlim dependency
`SemaphoreSlim` did not guarantee FIFO ordering, which we had mistakenly taken for granted for years. To date, no violation of this incorrect assumption has been detected, but we want to make a strong guarantee of FIFO ordering so folks can use `ReentrantSemaphore` (which is based on `AsyncSemaphore`) as a JTF-aware queue.