Add `AwaitExtensions.ConfigureAwaitForAggregateException` extension method
This allows awaiting a `Task` without losing all but the first of the `InnerExceptions` when it faults.

Besides adding tests for it, I modify an existing test to *use* it since it was failing with a timeout on occasion and I suspect the reason is found in another of the `Task.WhenAll` arguments' failures.