Move a condition check inside RemoveDependingSynchronousTask to one of its callers.
That condition check is not necessary during cleanup when a dependency is removed.  When ref-count of completed syncTask is left behind due to the earlier error (mismatching logic), this one can prevent the left behind refcount to be cleaned up, and will leave it behind even when the circular loop is removed. Due to the new logic skip rescanning when it thinks no circular reference can be formed more aggresively, it will remain in the graph for long time without being cleaned up.