Fix a task continuation hang related to project lock.
We noticed web project code gets a project property value (through a rule), then fires an internal event. One of the event handler JTF.Run and switches to the UI thread. It blocks other project code happens to read the project state at the same time due to the special task continuation deadlock.

The change here takes advantage of the logic in the task libraray, which spins off async continuation first, then calls other continuations in sequence. The logic here is that we join existing computation in async way to prevent this type of deadlock, and let the first continuation keeps the original behavior to reduce potential perf impact.