Fix performance problems when we have lots of tasks waiting on AsyncLazy<T>.
The problem is that ListOfOftenOne was not designed for the current usage. When we add lots of items to it, it will repeat allocating arrays. A new data structure was created to retain some advantage of the original one like more efficient to hold 0 or 1 items. (And comparing to List<T>, it saves memory for small amount of items, as we do not need any thread-safe related design here.

We don't allow enumerating items directly so we don't need any reentry protection. It was completely designed to fit the current need. Consider we rarely remove a JoinableTask from a collection, and most JoinAsync will wait until the task to finish, it is rarely that we need remove any added item, so the design is to keep Add close to O(1), and remove O(n).