Fix a problem that reader/writer lock hardcoded task scheduler.
Preparing resource is a common routine to access evaluated project, most of the time, it is close to no-op when the evaluation is up to date. Because the code always schedules the call to the thread pool, it defeats the very reason to have a priority queue
for high priority work when thread pool is too busy. It also makes it hard to throttle project evaluations through special task scheduler.

The change is to try not to change the behavior, when the code is called on UI thread (as it can lead into new problems), but try to use the current task scheduler if possible.