reduce overhead when running in no main thread mode (#1502)
* Do not track nested factory when there is no main thread

Generally, post to context would not be used as SwitchToMainThread is no-op in this context. Even it is being called under certain corner cases, it would just queue the same work item to the thread pool multiple times, so better to prevent the allocation in the first place.

* Do not turn on SynchronouslyBlockingMainThread flag on when the JTF is used in no main thread mode.

Because IsOnMainThread is just to check the thread id, this can be turned on randomly when the code is running in the initial thread pool thread. It would turn on expensive dependency tracking graph, which is not what we expect to do.

* stop creating main thread synchronization context when running in no main thread mode.

---------

Co-authored-by: Lifeng Lu <lifengl@microsoft.com>